requirements.txt
uvicorn==0.18.3
fastapi==0.85.0
python_dotenv==0.21.0
python_jose==3.3.0
pymongo==4.3.3
python-multipart==0.0.5
passlib[bcrypt]